+++ Diversity/src/server/Operations.java
@@ -12,6 +12,7 @@ public class Operations {
 		op.put("load", 2);
 		op.put("globalsentiment", 3);
 		op.put("getposts", 4);
+		op.put("getproducts",5);
 		
 		
 	}